+++ channels/testing.py
@@ -5,10 +5,10 @@ from .channel import PipeChannel
 
 class TestChannel(PipeChannel):
 
-    def __init__(self):
+    def __init__(self, **kwargs):
         self._in_r, self._out_w = os.pipe()
         self._out_r, self._in_w = os.pipe()
-        super().__init__(self._out_r, self._out_w)
+        super().__init__(self._out_r, self._out_w, **kwargs)
         self._inner = PipeChannel(self._in_r, self._in_w)
 
     def get(self):