+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0] - 2018-10-25
 ### Added
 - `buffering` parameter for Poller class
 - `buffering` parameter for channels
@@ -31,4 +33,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - TestChannel class
 - Poller class
 
-[Unreleased]: https://github.com/aragaer/channels/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/aragaer/channels/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/aragaer/channels/compare/v0.1.1...v0.2.0